+++ docs/history.md
@@ -6,6 +6,8 @@
 ### _v0.2.2_ - October 14th, 2023
 
 - Fallen now needs to be split in order to accomplish its objective.
+- Gurgling now requires 2 of them to be alive in order to accomplish its objective.
+- Turdling now requires 2 of them to be alive in order to accomplish its objective.
 
 ### _v0.2.0_ - October 14th, 2023